+++ package.json
@@ -24,6 +24,7 @@
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.462.0",
     "next": "15.0.3",
+    "next-auth": "^5.0.0-beta.25",
     "next-themes": "^0.4.3",
     "react": "19.0.0-rc-66855b96-20241106",
     "react-dom": "19.0.0-rc-66855b96-20241106",
